Moved all services to their own folder
